rebalancing more tests
